IoTApp/requirements.txt
Flask==1.1.1
Flask-WTF==0.14.2
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.10.3
paho-mqtt==1.5.0
MarkupSafe==1.1.1
Werkzeug==0.16.0
wheel==0.24.0
